minor fixes

+++ rust/services/compute-api/src/features/handlers.rs
@@ -182,7 +182,7 @@ pub async fn get_deployment_handler(
     State(mut redis): State<Redis>,
 ) -> Result<impl IntoResponse, AppError> {
     let user_id: Uuid = claims.sub;
-    let count = minutes * 60 / cfg.prometheus.scrape_interval;
+    let count = (minutes * 60 / cfg.prometheus.scrape_interval) as isize;
 
     let deployment =
         DeploymentRepository::get_by_id(&user_id, &deployment_id, &database.pool).await?;
@@ -207,7 +207,7 @@ pub async fn get_deployments_handler(
 ) -> Result<impl IntoResponse, AppError> {
     let user_id: Uuid = claims.sub;
     // scrape_interval in seconds, so we convert minutes query to seconds
-    let count = minutes * 60 / cfg.prometheus.scrape_interval;
+    let count = (minutes * 60 / cfg.prometheus.scrape_interval) as isize;
 
     let (deployments, total) =
         DeploymentRepository::get_all_by_project(&user_id, &project_id, &p, &database.pool).await?;
